refactor: split ParallelGroupRenderer into smaller components
- Break down ParallelGroupRenderer.tsx (200  88 lines)
- Extract useParallelGroupLogic.ts (~60 lines) - business logic
- Create ParallelGroupCountInput.tsx (47 lines) - count input field
- Create ParallelGroupContent.tsx (57 lines) - reusable content renderer
- Create ParallelGroupTabs.tsx (56 lines) - tabs mode display
- Create ParallelGroupSequential.tsx (37 lines) - sequential mode

Improve: Single Responsibility Principle applied
Improve: Better code organization and maintainability
Improve: Reusable components for tabs/sequential modes
Fix: Maintain exact same functionality and UI
Fix: All TypeScript types preserved

Breaking changes: NONE
Visual changes: NONE
